requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy
opencv-python
streamlit
pillow
matplotlib
python-docx